checker-backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
billiard==4.2.3
celery==5.5.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
flake8==7.3.0
gunicorn==23.0.0
iniconfig==2.3.0
kombu==5.5.4
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycodestyle==2.14.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
